% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePerFeatureDF.R
\name{makePerFeatureDF}
\alias{makePerFeatureDF}
\alias{makePerFeatureDF.SummarizedExperiment}
\alias{makePerFeatureDF.SingleCellExperiment}
\alias{makePerFeatureDF.ExpressionSet}
\alias{makePerFeatureDF.Seurat}
\title{Create a per-feature data.frame}
\usage{
makePerFeatureDF(x, ...)

\method{makePerFeatureDF}{SummarizedExperiment}(
  x,
  features = TRUE,
  assay = NULL,
  swap_rownames = NULL,
  use_rowdata = TRUE,
  ...,
  check_names = FALSE
)

\method{makePerFeatureDF}{SingleCellExperiment}(
  x,
  features = TRUE,
  assay = NULL,
  swap_rownames = NULL,
  use_rowdata = TRUE,
  ...,
  use_altexps = FALSE,
  prefix_altexps = FALSE,
  check_names = FALSE
)

\method{makePerFeatureDF}{ExpressionSet}(
  x,
  features = TRUE,
  assay = NULL,
  swap_rownames = NULL,
  use_rowdata = TRUE,
  ...,
  check_names = FALSE
)

\method{makePerFeatureDF}{Seurat}(
  x,
  features = TRUE,
  assay = NULL,
  layer = NULL,
  swap_rownames = NULL,
  use_rowdata = TRUE,
  ...,
  check_names = FALSE
)
}
\arguments{
\item{x}{A
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment},
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment},
\link[Biobase:class.ExpressionSet]{ExpressionSet} or
\link[SeuratObject:Seurat-class]{Seurat} object.}

\item{...}{Not used currently}

\item{features}{Logical scalar indicating whether feature assay data of \code{x}
should be included. Alternatively, Character or integer vector specifying the
features for which to extract expression profiles across samples.}

\item{assay}{String or integer scalar indicating the assay to use to obtain
expression values. Must refer to a matrix-like object with integer or numeric
values. If \code{NULL}, the default assay will be used:
\itemize{
\item \link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment}: the
\verb{1st} assay
\item \link[Biobase:class.ExpressionSet]{ExpressionSet}: assay from
\link[Biobase:exprs]{exprs}
\item \link[SeuratObject:Seurat-class]{Seurat}:
assay from \link[SeuratObject:DefaultAssay]{DefaultAssay}
}}

\item{swap_rownames}{String or integer specifying the feature metadata column
containing the features names. If \code{NULL}, \code{rownames(x)} will be used.}

\item{use_rowdata}{Logical scalar indicating whether feature metadata of \code{x}
should be included. Alternatively, a character or integer vector specifying
the feature metadata fields to use.}

\item{check_names}{A boolean value indicating whether column names of the
output should be made syntactically valid and unique.}

\item{use_altexps}{A boolean value indicating whether (meta)data should be
extracted for \link[SingleCellExperiment:altExps]{alternative} experiments in
\code{x}.}

\item{prefix_altexps}{A boolean value indicating whether
\link[SingleCellExperiment:altExps]{altExp}-derived fields should be prefixed
with the name of the alternative Experiment.}

\item{layer}{A string indicates the layer to get.}
}
\value{
A \link{data.frame}.
}
\description{
Create a per-feature data.frame (i.e., where each row represents a feature /
gene) from a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment}, most
typically for creating custom ggplot2 plots.
}
