% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{mock}
\alias{mock}
\alias{mockES}
\alias{mockSE}
\alias{mockSCE}
\alias{mockSeurat}
\title{Mock up sequencing counts data}
\usage{
mockES(ncells = 200, ngenes = 2000)

mockSE(ncells = 200, ngenes = 2000)

mockSCE(ncells = 200, ngenes = 2000, nspikes = 100)

mockSeurat(ncells = 200, ngenes = 2000)
}
\arguments{
\item{ncells}{An integer scalar, number of cells to simulate.}

\item{ngenes}{An integer scalar, number of genes to simulate.}

\item{nspikes}{An integer scalar, number of spike-in transcripts to simulate.
If \code{NULL}, no spike data will be added.}
}
\value{
\itemize{
\item \code{mockES}: A \link[Biobase:class.ExpressionSet]{ExpressionSet} object.
}

\itemize{
\item \code{mockSE}: A
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment} object.
}

\itemize{
\item \code{mockSCE}: A
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment} object.
}

\itemize{
\item \code{mockSeurat}: A \link[SeuratObject:Seurat-class]{Seurat} object.
}
}
\description{
Mock up a \link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment},
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment},
\link[Biobase:class.ExpressionSet]{ExpressionSet} or
\link[SeuratObject:Seurat-class]{Seurat} object containing simulated data.
}
\examples{
mockES()
mockSE()
mockSCE()
mockSeurat()
}
